Add support for self-signed certificates
If you are using self-signed certificates with a custom CA in your organisation, you need to supply the trusted root files to the trusted root CA store within alpine linux in order to make an SSL-encrypted connection. 

Disabling certificate verification with `VSPHERE_IGNORE_SSL=TRUE` is not viable in production environments.